Optimise and cleanup code
