+++ src/test/java/uk/gov/hmcts/opal/authentication/dao/AzureDaoTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.authentication.dao.impl;
+package uk.gov.hmcts.opal.authentication.dao;
 
 import com.nimbusds.oauth2.sdk.http.HTTPResponse;
 import org.apache.http.HttpStatus;
@@ -14,6 +14,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.opal.authentication.client.OAuthClient;
 import uk.gov.hmcts.opal.authentication.config.AuthConfigurationProperties;
 import uk.gov.hmcts.opal.authentication.config.AuthProviderConfigurationProperties;
+import uk.gov.hmcts.opal.authentication.dao.AzureDao;
 import uk.gov.hmcts.opal.authentication.exception.AzureDaoException;
 import uk.gov.hmcts.opal.authentication.model.OAuthProviderRawResponse;
 
@@ -25,7 +26,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class AzureDaoImplTest {
+class AzureDaoTest {
 
     @Mock
     private AuthConfigurationProperties authenticationConfiguration;
@@ -37,7 +38,7 @@ class AzureDaoImplTest {
     private OAuthClient azureActiveDirectoryB2CClient;
 
     @InjectMocks
-    private AzureDaoImpl azureDaoImpl;
+    private AzureDao azureDao;
 
 
     @Test
@@ -46,7 +47,7 @@ class AzureDaoImplTest {
         when(azureActiveDirectoryB2CClient.fetchAccessToken(any(), any(), any(), any(), any(), any())).thenReturn(
             response);
 
-        OAuthProviderRawResponse rawResponse = azureDaoImpl.fetchAccessToken(
+        OAuthProviderRawResponse rawResponse = azureDao.fetchAccessToken(
             "CODE",
             authenticationProviderConfiguration,
             authenticationConfiguration
@@ -64,7 +65,7 @@ class AzureDaoImplTest {
     @ValueSource(strings = {" "})
     @NullAndEmptySource
     void fetchAccessTokenShouldThrowExceptionWhenProvidedCodeIsBlankOrNull(String code) {
-        AzureDaoException exception = assertThrows(AzureDaoException.class, () -> azureDaoImpl.fetchAccessToken(
+        AzureDaoException exception = assertThrows(AzureDaoException.class, () -> azureDao.fetchAccessToken(
             code, authenticationProviderConfiguration, authenticationConfiguration));
 
         assertEquals("Null code not permitted", exception.getMessage());
@@ -78,7 +79,7 @@ class AzureDaoImplTest {
 
         AzureDaoException exception = assertThrows(
             AzureDaoException.class,
-            () -> azureDaoImpl.fetchAccessToken(
+            () -> azureDao.fetchAccessToken(
                 "CODE",
                 authenticationProviderConfiguration,
                 authenticationConfiguration
