+++ src/test/java/uk/gov/hmcts/opal/authentication/service/AuthenticationServiceTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.opal.authentication.service.impl;
+package uk.gov.hmcts.opal.authentication.service;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -28,7 +28,7 @@ import static org.mockito.ArgumentMatchers.notNull;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class AuthenticationServiceImplTest {
+class AuthenticationServiceTest {
 
     private static final URI DUMMY_AUTH_URI = URI.create("DUMMY_AUTH_URI");
     private static final URI DUMMY_LOGOUT_URI = URI.create("DUMMY_LOGOUT_URI");
@@ -37,7 +37,7 @@ class AuthenticationServiceImplTest {
     private static final String DUMMY_ID_TOKEN = "DUMMY ID TOKEN";
 
     @InjectMocks
-    private AuthenticationServiceImpl authenticationService;
+    private AuthenticationService authenticationService;
 
     @Mock
     private TokenValidator tokenValidator;
